changed from custom pspp and gtk-mac-bundler packages to default
The custom pspp package was used because the reloc variant was not there.
This has been changed in macport, so now the default package can be used.

The custom gtk-mac-bundler package was required because 0.7.4 plus some
non released git changes was required. In addition the examples were not
adapted to macports and not installed. osxbundler relies on the
gtk3-launcher.sh from the examples directory. This is now changed in
macports, so no custom package is required anymore.