added version information to bundler and build
With this patch the git base build will be build with make VERSION=XXXX
where the version is based on the release version and the git version as
in Bens snapshots of pspp: http://benpfaff.org/~blp/pspp-master/

If the bundle is based on the version in macports, then the macports version
is taken.

The version information is adapted in the .plist file.