fixed folder access problem on Catalina and following - bundle version 1->2
With the bundle version 1.4.1-1 you cannot access the Desktop, Downloads
and other directories. The reason is that pspp is called indirectly via
a shell script. MacOS is then confused and cannot track the access rights
of the application.

I removed the shell script and introduced the link in the MacOS bundle
directory again. Works on Catalina and Monterey.

I increased the bundle version from 1 to 2 and changed the app id
from org.gtk.pspp org.fsf.pspp.